requirements.txt
flask
flask-smorest
python-dotenv
flask-sqlalchemy
spacy
